There can now only be 512 (2^9) uses of a __COUNTER__ per file per translation unit, though with collisions in the 2^23 bits of file hashes, it's possible that the number is significantly reduced

+++ lib/Compile/PatchedMacroTracker.h
@@ -111,7 +111,7 @@ class PatchedMacroTracker : public clang::PPCallbacks {
   // semblance of uniqueness, but also we want to try to have better
   // reproducibility across translation units, where the same macro in the
   // same header using a `__COUNTER__` ends up expanding to the same value.
-  std::unordered_map<uint16_t, unsigned> next_counter_value;
+  std::unordered_map<unsigned, unsigned> next_counter_value;
   uint16_t last_counter_id{0};
 
  public:
